added shape img to about us, and stylized the creator cards.
