+++ src/assets/aboutUs.js
@@ -1,0 +1,45 @@
+export const aboutUs = [
+  {
+    name: 'forest',
+    linkedIn: 'https://www.linkedin.com/in/forestheims/',
+    gitHub: 'https://github.com/forestheims',
+    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
+    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
+    qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng] velit,
+ `,
+    photo: 'https://cdn-icons-png.flaticon.com/128/6080/6080922.png',
+    geoTone: './user/:username',
+  },
+  {
+    name: 'Michelle',
+    linkedIn: 'https://www.linkedin.com/in/michellenygren/',
+    gitHub: 'https://github.com/michellerenehey',
+    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
+    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
+    qui dolorem ipsum, quia dolor sit amet consectetur a`,
+    photo: 'https://cdn-icons-png.flaticon.com/128/5853/5853921.png',
+    geoTone: './user/:username',
+  },
+  {
+    name: 'Jordan',
+    linkedIn: 'https://www.linkedin.com/in/jlaurentpdx/',
+    gitHub: 'https://github.com/jlaurentpdx',
+    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
+    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
+    qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng] velit,
+  `,
+    photo: 'https://cdn-icons-png.flaticon.com/128/136/136830.png',
+    geoTone: './user/:username',
+  },
+  {
+    name: 'Minoka',
+    linkedIn: 'https://www.linkedin.com/in/minoka-kakizaki/',
+    gitHub: 'https://github.com/kakizaki55',
+    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
+    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
+    qui dolorem ipsum, q`,
+    photo:
+      'https://cdn-icons.flaticon.com/png/128/697/premium/697818.png?token=exp=1650407018~hmac=6cdd68eb78b4307fc410a164113112fb',
+    geoTone: './user/:username',
+  },
+];
