+++ src/views/About/About.css
@@ -2,13 +2,14 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
+  justify-content: space-between;
+  width: 200px;
   padding: 10px;
 }
 
 .aboutContainer {
   display: flex;
+  flex-wrap: wrap;
   gap: 45px;
 }
 
@@ -19,5 +20,8 @@
 
 .aboutCard a img:hover {
   cursor: pointer;
+  margin: 15px;
   width: 55px;
+  border-radius: 100%;
+  background-color: aqua;
 }
