+++ src/views/About/About.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './About.css';
-import { aboutUs } from './aboutUs.js';
+import { aboutUs } from '../../assets/aboutUs.js';
 import githubLogo from '../../assets/github.png';
 import linkedInLogo from '../../assets/linkedin.png';
 
@@ -10,8 +10,8 @@ export default function About() {
     <div className={styles.aboutContainer}>
       {aboutUs.map((each) => (
         <div key={each.name} className={styles.aboutCard}>
+          <img src={each.photo} />
           <h2>{each.name}</h2>
-          {/* <img src={each.photo} /> */}
           <span>{each.bio}</span>
           <div>
             <a href={each.gitHub}>
