--- src/views/About/aboutUs.js
@@ -1,58 +1,0 @@
-export const aboutUs = [
-  {
-    name: 'forest',
-    linkedIn: 'https://www.linkedin.com/in/forestheims/',
-    gitHub: 'https://github.com/forestheims',
-    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
-    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
-    qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng] velit,
-    sed quia non numquam [do] eius modi tempora inci[di]dunt, ut labore et
-    dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis
-    nostrum[d] exercitationem ullam corporis suscipit laboriosam, nisi ut
-    aliquid ex ea commodi consequatur? [D]Quis autem vel eum iure
-    reprehenderit, qui in ea volu`,
-    photo: '',
-  },
-  {
-    name: 'Michelle',
-    linkedIn: 'https://www.linkedin.com/in/michellenygren/',
-    gitHub: 'https://github.com/michellerenehey',
-    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
-    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
-    qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng] velit,
-    sed quia non numquam [do] eius modi tempora inci[di]dunt, ut labore et
-    dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis
-    nostrum[d] exercitationem ullam corporis suscipit laboriosam, nisi ut
-    aliquid ex ea commodi consequatur? [D]Quis autem vel eum iure
-    reprehenderit, qui in ea volu`,
-    photo: '',
-  },
-  {
-    name: 'Jordan',
-    linkedIn: 'https://www.linkedin.com/in/jlaurentpdx/',
-    gitHub: 'https://github.com/jlaurentpdx',
-    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
-    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
-    qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng] velit,
-    sed quia non numquam [do] eius modi tempora inci[di]dunt, ut labore et
-    dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis
-    nostrum[d] exercitationem ullam corporis suscipit laboriosam, nisi ut
-    aliquid ex ea commodi consequatur? [D]Quis autem vel eum iure
-    reprehenderit, qui in ea volu`,
-    photo: '',
-  },
-  {
-    name: 'Minoka',
-    linkedIn: 'https://www.linkedin.com/in/minoka-kakizaki/',
-    gitHub: 'https://github.com/kakizaki55',
-    bio: `natur aut odit aut fugit, sed quia consequuntur magni dolores
-    eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est,
-    qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng] velit,
-    sed quia non numquam [do] eius modi tempora inci[di]dunt, ut labore et
-    dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis
-    nostrum[d] exercitationem ullam corporis suscipit laboriosam, nisi ut
-    aliquid ex ea commodi consequatur? [D]Quis autem vel eum iure
-    reprehenderit, qui in ea volu`,
-    photo: '',
-  },
-];
